+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-Version 0.7 *(2014-01-28)*
+Version 0.7 *(2014-01-30)*
 ----------------------------
   * API break: MortarActivityScope#onResume is gone, and as you might expect
     Bundler#onLoad is not called at resume time. It just wasn't useful. See
@@ -9,6 +9,10 @@ Version 0.7 *(2014-01-28)*
 
   * API break: Presenter is no longer a Bundler, and its onLoad method
     is never called with a null view.
+
+  * New: Mortar#createRootScope(boolean) for simpler root scope creation.
+
+  * New: Hello Mortar sample app.
 
 Version 0.6 *(2014-01-20)*
 ----------------------------